move draw functions to draw.py and switch to block
