+++ main.py
@@ -1,134 +1,10 @@
 import pygame
-from pygame import Color, Surface, Vector2
+from pygame import Vector2
 
 import block as B
-
-# COLORS
-WHITE = Color(255, 255, 255)
-LIGHTGRAY = Color(160, 160, 160)
-GRAY = Color(64, 64, 64)
-BLACK = Color(0, 0, 0)
-RED = Color(255, 0, 0)
-YELLOW = Color(255, 255, 0)
-GREEN = Color(0, 255, 0)
-CYAN = Color(0, 255, 255)
-BLUE = Color(0, 0, 255)
+from draw import *
 
 MOVEMENT_DELAY = 0.5
-
-Y_HAT = Vector2(-50, 25)
-X_HAT = Vector2(50, 25)
-
-
-def drawIsometricTop(
-    surface: Surface,
-    color: Color,
-    center: Vector2,
-    width: int = 100,
-    height: int = 50,
-    outline: bool = False,
-    outlineColor: Color = BLACK,
-) -> None:
-    """draws a top facing square face in an isometric view"""
-    p1 = Vector2(center.x - width / 2, center.y)
-    p2 = Vector2(center.x, center.y + height / 2)
-    p3 = Vector2(center.x + width / 2, center.y)
-    p4 = Vector2(center.x, center.y - height / 2)
-    coordinates = [p1, p2, p3, p4]
-    pygame.draw.polygon(surface, color, coordinates)
-    if outline:
-        pygame.draw.line(surface, outlineColor, p1, p2)
-        pygame.draw.line(surface, outlineColor, p3, p2)
-        pygame.draw.line(surface, outlineColor, p3, p4)
-        pygame.draw.line(surface, outlineColor, p1, p4)
-
-
-def drawIsometricRight(
-    surface: Surface,
-    color: Color,
-    center: Vector2,
-    width: int = 100,
-    height: int = 50,
-    outline: bool = False,
-    outlineColor: Color = BLACK,
-) -> None:
-    """draws a right facing square face in an isometric view"""
-    p1 = Vector2(center.x - width / 4, center.y + height / 4 - height / 2)
-    p2 = Vector2(center.x - width / 4, center.y + height / 4 + height / 2)
-    p3 = Vector2(center.x + width / 4, center.y - height / 4 + height / 2)
-    p4 = Vector2(center.x + width / 4, center.y - height / 4 - height / 2)
-    coordinates = [p1, p2, p3, p4]
-    pygame.draw.polygon(surface, color, coordinates)
-    if outline:
-        pygame.draw.line(surface, outlineColor, p1, p2)
-        pygame.draw.line(surface, outlineColor, p3, p2)
-        pygame.draw.line(surface, outlineColor, p3, p4)
-        pygame.draw.line(surface, outlineColor, p1, p4)
-
-
-def drawIsometricLeft(
-    surface: Surface,
-    color: Color,
-    center: Vector2,
-    width: int = 100,
-    height: int = 50,
-    outline: bool = False,
-    outlineColor: Color = BLACK,
-) -> None:
-    """draws a left facing square face in an isometric view"""
-    p1 = Vector2(center.x + width / 4, center.y + height / 4 - height / 2)
-    p2 = Vector2(center.x + width / 4, center.y + height / 4 + height / 2)
-    p3 = Vector2(center.x - width / 4, center.y - height / 4 + height / 2)
-    p4 = Vector2(center.x - width / 4, center.y - height / 4 - height / 2)
-    coordinates = [p1, p2, p3, p4]
-    pygame.draw.polygon(surface, color, coordinates)
-    if outline:
-        pygame.draw.line(surface, outlineColor, p1, p2)
-        pygame.draw.line(surface, outlineColor, p3, p2)
-        pygame.draw.line(surface, outlineColor, p3, p4)
-        pygame.draw.line(surface, outlineColor, p1, p4)
-
-
-def drawCube(
-    surface: Surface,
-    color: Color,
-    center: Vector2,
-    sideLength: int,
-    outline: bool = True,
-    outlineColor: Color = BLACK,
-) -> None:
-    """draws an isometric cube"""
-    drawIsometricTop(
-        surface=surface,
-        color=color,
-        center=Vector2(center.x, center.y - sideLength / 2),
-        width=sideLength * 2,
-        height=sideLength,
-        outline=outline,
-        outlineColor=outlineColor,
-    )
-    drawIsometricRight(
-        surface=surface,
-        color=color,
-        center=Vector2(center.x + sideLength / 2, center.y + sideLength / 4),
-        width=sideLength * 2,
-        height=sideLength,
-        outline=outline,
-        outlineColor=outlineColor,
-    )
-    drawIsometricLeft(
-        surface=surface,
-        color=color,
-        center=Vector2(center.x - sideLength / 2, center.y + sideLength / 4),
-        width=sideLength * 2,
-        height=sideLength,
-        outline=outline,
-        outlineColor=outlineColor,
-    )
-
-
-def CoordToScreen(Coord: Vector2) -> Vector2:
-    return Coord.x * X_HAT + Coord.y * Y_HAT
 
 
 def main():
@@ -141,7 +17,7 @@ def main():
     running = True
     dt = 0
     # load level
-    cubePos = Vector2(3, 1)
+    block = B.Block(Vector2(5, 3))
     movementTimeout = 0
     # game loop
     while running:
@@ -156,23 +32,23 @@ def main():
 
         if movementTimeout < 0:
             if keys[pygame.K_d]:
-                cubePos += Vector2(1, 0)
+                block._pos += Vector2(1, 0)
                 movementTimeout = MOVEMENT_DELAY
             if keys[pygame.K_a]:
-                cubePos -= Vector2(1, 0)
+                block._pos += Vector2(-1, 0)
                 movementTimeout = MOVEMENT_DELAY
 
             if keys[pygame.K_w]:
-                cubePos -= Vector2(0, 1)
+                block._pos += Vector2(0, -1)
                 movementTimeout = MOVEMENT_DELAY
             if keys[pygame.K_s]:
-                cubePos += Vector2(0, 1)
+                block._pos += Vector2(0, 1)
                 movementTimeout = MOVEMENT_DELAY
         #    check win/fail conditions
 
         #    animate
         screen.fill(GRAY)
-        drawCube(screen, RED, CoordToScreen(cubePos), 50)
+        block.draw(screen)
         pygame.display.flip()
 
         movementTimeout -= dt
